미사일 시간에 따라서 사용할 수 있게 연구중

+++ Robotopia/Classes/MissilePlayerMelee.h
@@ -35,10 +35,10 @@ public:
 	CREATE_FUNC(MissilePlayerMelee);
 
 private:
-	long	m_Duration;
+	float	m_Duration;
 	bool	m_IsPrevTimeCheck;
-	long	m_PrevTime;
-	long	m_CurTime;
+	float	m_PrevTime;
+	float 	m_CurTime;
 
 };
